back off yesterday's change -- won't compile on sun and will break mspew for mode.eq.2.  needs thought!!

+++ maplib.f
@@ -567,7 +567,7 @@ C     .. Scalar Arguments ..
       INTEGER IUNIT
 C     ..
 C     .. Array Arguments ..
-      BYTE X(*)
+      REAL X(*)
 C     ..
 C     .. Scalars in Common ..
       REAL AMAX,AMEAN,AMIN,ARMS
